Using Graphics2D class
There are things you can do with a Graphics2D reference that you can’t do with a Graphics reference. The object referenced by the ‘g’ parameter is actually an instance of the Graphics2D class but dye to runtime polymorphism you can't directly call a method of that class. If you need to use a method from the Graphics2D class, you can’t use the the paintComponent parameter (‘g’) straight from the method. But you can cast it with a new Graphics2D variable.